pyproject.toml
[tool.poetry]
name = "forecasting-tools"
version = "0.1.0"
description = "AI forecasting and research tools to help humans reason about and forecast the future"
authors = ["Benjamin Wilson <mokoresearch@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

# Core packages
langchain-community = "0.3.4"
langchain-openai = "0.2.2"
langchain-core = "0.3.9"
langchain-anthropic = "0.2.4"
openai = "1.51.0"

# Utility packages
aiofiles = "^23.2.1"
aiohttp = "^3.9.3"
aiolimiter = "^1.1.0"
argcomplete = "^1.10.3"
asyncio = "^3.4.3"
bs4 = "^0.0.1"
nest-asyncio = "^1.5.8"
numpy = "^1.26.0"
pipreqs = "^0.4.13"
pydantic = "^2.9.2"
python-dotenv = "^1.0.0"
regex = "^2023.8.8"
requests = "^2.32.3"
tiktoken = "^0.8.0"
streamlit = "^1.36.0"
tenacity = "^8.5.0"
pillow = "^10.4.0"
transformers = "^4.44.1"
typeguard = "^4.3.0"

[tool.poetry.group.dev.dependencies]
scipy = "^1.12.0"
pytest = "^8.3.2"
pytest-xdist = "^3.3.1"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
